Add decompilation and split code
